CaseViewSet has new filters and ordering and broken up case views to give more overview
