+++ app/apps/cases/views/bwv.py
@@ -1,11 +1,9 @@
 import datetime
 import logging
-import mimetypes
-import os
 from collections import OrderedDict
 
 import requests
-from apps.addresses.utils import search
+from apps.cases.forms import ImportBWVCaseDataForm
 from apps.cases.models import (
     Case,
     CaseClose,
@@ -36,24 +34,6 @@ from apps.cases.serializers import (
     StartWorkflowSerializer,
     SubjectSerializer,
 )
-from apps.cases.swagger_parameters import date as date_parameter
-from apps.cases.swagger_parameters import no_pagination as no_pagination_parameter
-from apps.cases.swagger_parameters import open_cases as open_cases_parameter
-from apps.cases.swagger_parameters import open_status as open_status_parameter
-from apps.cases.swagger_parameters import postal_code as postal_code_parameter
-from apps.cases.swagger_parameters import reason as reason_parameter
-from apps.cases.swagger_parameters import start_date as start_date_parameter
-from apps.cases.swagger_parameters import street_name as street_name_parameter
-from apps.cases.swagger_parameters import street_number as street_number_parameter
-from apps.cases.swagger_parameters import suffix as suffix_parameter
-from apps.cases.swagger_parameters import theme as theme_parameter
-from apps.cases.swagger_parameters import ton_ids as ton_ids_parameter
-from apps.debriefings.models import Debriefing
-from apps.debriefings.serializers import ViolationTypeSerializer
-from apps.decisions.serializers import DecisionTypeSerializer
-from apps.events.mixins import CaseEventsMixin
-from apps.schedules.serializers import ThemeScheduleTypesSerializer
-from apps.summons.serializers import SummonTypeSerializer
 from apps.users.models import User
 from apps.users.permissions import (
     CanCloseCase,
@@ -62,571 +42,24 @@ from apps.users.permissions import (
 )
 from apps.visits.models import Visit
 from apps.visits.serializers import VisitSerializer
-from apps.workflow.models import CaseWorkflow, GenericCompletedTask, WorkflowOption
+from apps.workflow.models import GenericCompletedTask
 from apps.workflow.serializers import (
     GenericCompletedTaskSerializer,
     WorkflowOptionSerializer,
 )
 from django.conf import settings
-from django.contrib.auth import get_user_model
-from django.contrib.auth.decorators import user_passes_test
 from django.contrib.auth.mixins import UserPassesTestMixin
-from django.http import HttpResponse, HttpResponseBadRequest
-from django.shortcuts import get_object_or_404, redirect
+from django.shortcuts import redirect
 from django.urls import reverse
 from django.views.generic.edit import FormView
-from drf_spectacular.utils import extend_schema
-from rest_framework import mixins, status, viewsets
-from rest_framework.decorators import action
-from rest_framework.exceptions import APIException
 from rest_framework.generics import (
     ListAPIView,
     ListCreateAPIView,
     RetrieveUpdateDestroyAPIView,
 )
-from rest_framework.pagination import LimitOffsetPagination
-from rest_framework.permissions import SAFE_METHODS
-from rest_framework.response import Response
 from utils.api_queries_bag import do_bag_search_address_exact
 
-from .forms import ImportBWVCaseDataForm
-
 logger = logging.getLogger(__name__)
-
-
-class CaseStateViewSet(viewsets.ViewSet):
-    """
-    Pushes the case state
-    """
-
-    permission_classes = rest_permission_classes_for_top()
-    serializer_class = CaseStateSerializer
-    queryset = CaseState.objects.all()
-
-    @action(
-        detail=True,
-        url_path="update-from-top",
-        methods=["post"],
-        serializer_class=PushCaseStateSerializer,
-    )
-    def update_from_top(self, request, pk):
-        logger.info("Receiving pushed state update")
-        data = request.data
-        serializer = self.serializer_class(data=data)
-
-        if not serializer.is_valid():
-            logger.error("Serializer error: {serializer.errors}")
-            raise APIException(f"Serializer error: {serializer.errors}")
-
-        try:
-            case_state = CaseState.objects.get(id=pk)
-            case_state.users.clear()
-            user_emails = data.get("user_emails", [])
-            logger.info(f"Updating CaseState {len(user_emails)} users")
-            user_model = get_user_model()
-
-            for user_email in user_emails:
-                user_object, _ = user_model.objects.get_or_create(email=user_email)
-                case_state.users.add(user_object)
-                logger.info("Added user to CaseState")
-
-            return Response(CaseStateSerializer(case_state).data)
-        except Exception as e:
-            logger.error(f"Could not process push data: {e}")
-            raise logger(f"Could not push data: {e}")
-
-
-class CaseViewSet(
-    CaseEventsMixin,
-    mixins.CreateModelMixin,
-    mixins.UpdateModelMixin,
-    mixins.ListModelMixin,
-    mixins.RetrieveModelMixin,
-    viewsets.GenericViewSet,
-):
-    permission_classes = rest_permission_classes_for_top()
-    serializer_class = CaseSerializer
-    queryset = Case.objects.all()
-
-    def get_permissions(self):
-        if self.action == "create" and self.request.method not in SAFE_METHODS:
-            self.permission_classes.append(CanCreateCase)
-        return super(CaseViewSet, self).get_permissions()
-
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        case = self.perform_create(serializer)
-
-        citizen_report_data = {"case": case.id}
-        citizen_report_data.update(request.data)
-        citizen_report_serializer = CitizenReportSerializer(
-            data=citizen_report_data,
-            context={"request": request},
-        )
-        if citizen_report_serializer.is_valid():
-            citizen_report_serializer.save()
-
-        headers = self.get_success_headers(serializer.data)
-        return Response(
-            serializer.data, status=status.HTTP_201_CREATED, headers=headers
-        )
-
-    def perform_create(self, serializer):
-        return serializer.save()
-
-    def get_serializer_class(self, *args, **kwargs):
-        if self.action in ["create", "update"]:
-            return CaseCreateUpdateSerializer
-
-        return self.serializer_class
-
-    @extend_schema(
-        parameters=[
-            date_parameter,
-            start_date_parameter,
-            open_cases_parameter,
-            theme_parameter,
-            reason_parameter,
-            open_status_parameter,
-            no_pagination_parameter,
-        ],
-        description="Case filter query parameters",
-        responses={200: CaseSerializer(many=True)},
-    )
-    def list(self, request):
-        date = request.GET.get(date_parameter.name, None)
-        start_date = request.GET.get(start_date_parameter.name, None)
-        open_cases = request.GET.get(open_cases_parameter.name, None)
-        theme = request.GET.get(theme_parameter.name, None)
-        reason = request.GET.get(reason_parameter.name, None)
-        open_status = request.GET.get(open_status_parameter.name, None)
-        no_pagination = request.GET.get(no_pagination_parameter.name, None)
-
-        queryset = self.get_queryset()
-
-        if date:
-            queryset = queryset.filter(start_date=date)
-        if start_date:
-            queryset = queryset.filter(start_date__gte=start_date)
-        if open_cases:
-            open_cases = open_cases == "true"
-            queryset = queryset.filter(end_date__isnull=open_cases)
-        if theme:
-            queryset = queryset.filter(theme=theme)
-        if reason:
-            queryset = queryset.filter(reason=reason)
-        if open_status:
-            queryset = queryset.filter(
-                case_states__end_date__isnull=True,
-                case_states__status__id__in=open_status.split(","),
-            ).distinct()
-
-        if no_pagination == "true":
-            serializer = CaseSerializer(queryset, many=True)
-            return Response(serializer.data)
-
-        paginator = LimitOffsetPagination()
-        context = paginator.paginate_queryset(queryset, request)
-        serializer = CaseSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        parameters=[
-            postal_code_parameter,
-            street_number_parameter,
-            street_name_parameter,
-            suffix_parameter,
-            theme_parameter,
-            ton_ids_parameter,
-        ],
-        description="Search query parameters",
-        responses={200: CaseSerializer(many=True)},
-        operation=None,
-    )
-    @action(detail=False, methods=["get"], url_path="search")
-    def search(self, request):
-        postal_code = request.GET.get(postal_code_parameter.name, None)
-        street_name = request.GET.get(street_name_parameter.name, None)
-        number = request.GET.get(street_number_parameter.name, None)
-        suffix = request.GET.get(suffix_parameter.name, None)
-        theme = request.GET.get(theme_parameter.name, None)
-        ton_ids = request.GET.get(ton_ids_parameter.name, None)
-
-        if postal_code is None and street_name is None:
-            return HttpResponseBadRequest(
-                "A postal_code or street_name queryparameter should be provided"
-            )
-        if postal_code is not None and number is None:
-            return HttpResponseBadRequest("number queryparameter is required")
-        if street_name is not None and number is None:
-            return HttpResponseBadRequest("number queryparameter is required")
-
-        address_queryset = search(
-            street_name=street_name,
-            postal_code=postal_code,
-            number=number,
-            suffix=suffix,
-        )
-
-        cases = Case.objects.none()
-        for address in address_queryset:
-            cases = cases | address.cases.all()
-
-        cases = cases.filter(end_date=None)
-
-        if theme:
-            cases = cases.filter(theme=theme)
-
-        if ton_ids:
-            cases = cases.filter(ton_ids__overlap=ton_ids.split(","))
-
-        paginator = LimitOffsetPagination()
-        context = paginator.paginate_queryset(cases, request)
-        serializer = CaseSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Get tasks for this Case",
-        responses={status.HTTP_200_OK: CaseWorkflowSerializer(many=True)},
-    )
-    @action(detail=True, methods=["get"], url_path="tasks")
-    def get_tasks(self, request, pk):
-        case = self.get_object()
-        request.user
-        queryset = CaseWorkflow.objects.filter(
-            case=case, tasks__isnull=False, tasks__completed=False
-        ).distinct()
-        paginator = LimitOffsetPagination()
-        context = paginator.paginate_queryset(queryset, request)
-        serializer = CaseWorkflowSerializer(
-            context, many=True, context={"request": request}
-        )
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Gets the Subjects associated with the given theme",
-        responses={status.HTTP_200_OK: SubjectSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="subjects",
-        methods=["get"],
-    )
-    def subjects(self, request, pk):
-        paginator = LimitOffsetPagination()
-        case = self.get_object()
-        query_set = case.subjects.all()
-
-        context = paginator.paginate_queryset(query_set, request)
-        serializer = SubjectSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Get WorkflowOption for this Case",
-        responses={status.HTTP_200_OK: WorkflowOptionSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="processes",
-        methods=["get"],
-        serializer_class=WorkflowOptionSerializer,
-    )
-    def get_workflow_options(self, request, pk):
-        """
-        Get WorkflowOption instances for this case. Currently this case detail linking
-        does not do anything. This is future proofing this rest call so that we can
-        show and not show processes based on the current state of the case
-        (for example not show the summon/aanschrijving process when we are in visit state)
-        """
-        case = get_object_or_404(Case, pk=pk)
-        serializer = WorkflowOptionSerializer(
-            WorkflowOption.objects.filter(theme=case.theme), many=True
-        )
-        return Response(serializer.data)
-
-    @extend_schema(
-        description="Start based on a WorkflowOption",
-    )
-    @action(
-        detail=True,
-        url_path="processes/start",
-        methods=["post"],
-        serializer_class=StartWorkflowSerializer,
-    )
-    def start_process(self, request, pk):
-        serializer = self.serializer_class(data=request.data)
-
-        if serializer.is_valid():
-            data = serializer.validated_data
-            case = self.get_object()
-            instance = data["workflow_option_id"]
-
-            workflow_type = CaseWorkflow.WORKFLOW_TYPE_SUB
-            if instance.to_directing_proccess:
-                workflow_type = CaseWorkflow.WORKFLOW_TYPE_DIRECTOR
-
-            CaseWorkflow.objects.create(
-                case=case,
-                workflow_type=workflow_type,
-                workflow_message_name=instance.message_name,
-            )
-
-            return Response(
-                data=f"Workflow has started {str(instance)}",
-                status=status.HTTP_200_OK,
-            )
-
-        return Response(
-            data="Workflow has not started. serializer not valid",
-            status=status.HTTP_500_INTERNAL_SERVER_ERROR,
-        )
-
-    @extend_schema(
-        description="Create citizen report instance associated with this case",
-        responses={status.HTTP_200_OK: CitizenReportSerializer()},
-    )
-    @action(
-        detail=True,
-        url_path="citizen-reports",
-        methods=["post"],
-        serializer_class=CitizenReportSerializer,
-    )
-    def citizen_reports(self, request, pk):
-        serializer = self.serializer_class(
-            data=request.data,
-            context={"request": request},
-        )
-        if serializer.is_valid():
-            data = serializer.validated_data
-
-            citizen_report = CitizenReport(**data)
-            citizen_report.save()
-            return Response(
-                data="CitizenReport added",
-                status=status.HTTP_200_OK,
-            )
-        return Response(
-            data="CitizenReport error. serializer not valid",
-            status=status.HTTP_500_INTERNAL_SERVER_ERROR,
-        )
-
-
-class CaseThemeViewSet(ListAPIView, viewsets.ViewSet):
-    permission_classes = rest_permission_classes_for_top()
-    serializer_class = CaseThemeSerializer
-    queryset = CaseTheme.objects.all()
-
-    @extend_schema(
-        description="Gets the reasons associated with the requested theme",
-        responses={status.HTTP_200_OK: CaseReasonSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="reasons",
-        methods=["get"],
-    )
-    def reasons(self, request, pk):
-        paginator = LimitOffsetPagination()
-        theme = self.get_object()
-        query_set = theme.reasons.all()
-
-        context = paginator.paginate_queryset(query_set, request)
-        serializer = CaseReasonSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Gets the SummonTypes associated with the given theme",
-        responses={status.HTTP_200_OK: SummonTypeSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="summon-types",
-        methods=["get"],
-    )
-    def summon_types(self, request, pk):
-        paginator = LimitOffsetPagination()
-        theme = self.get_object()
-        query_set = theme.summon_types.all()
-
-        context = paginator.paginate_queryset(query_set, request)
-        serializer = SummonTypeSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Gets the DecisionTypes associated with the given theme",
-        responses={status.HTTP_200_OK: DecisionTypeSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="decision-types",
-        methods=["get"],
-    )
-    def decision_types(self, request, pk):
-        paginator = LimitOffsetPagination()
-        theme = self.get_object()
-        query_set = theme.decision_types.all()
-
-        context = paginator.paginate_queryset(query_set, request)
-        serializer = DecisionTypeSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Gets the Scheduling Types associated with the given theme",
-        responses={status.HTTP_200_OK: ThemeScheduleTypesSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="schedule-types",
-        methods=["get"],
-    )
-    def schedule_types(self, request, pk):
-        theme = self.get_object()
-        serializer = ThemeScheduleTypesSerializer(theme)
-        return Response(serializer.data)
-
-    @extend_schema(
-        description="Gets the CaseStateTypes associated with the given theme",
-        responses={status.HTTP_200_OK: CaseStateTypeSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="state-types",
-        methods=["get"],
-    )
-    def state_types(self, request, pk):
-        paginator = LimitOffsetPagination()
-        theme = self.get_object()
-        query_set = theme.state_types.all()
-        if request.GET.get("role") == "toezichthouder":
-            query_set = query_set.filter(
-                id__in=theme.case_state_types_top.values_list("id", flat=True)
-            )
-
-        context = paginator.paginate_queryset(query_set, request)
-        serializer = CaseStateTypeSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Gets the ViolationTypes",
-        responses={status.HTTP_200_OK: ViolationTypeSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="violation-types",
-        methods=["get"],
-    )
-    def violation_types(self, request, pk):
-        paginator = LimitOffsetPagination()
-        types = [{"key": t[0]} for t in Debriefing.VIOLATION_CHOICES]
-        context = paginator.paginate_queryset(types, request)
-        serializer = ViolationTypeSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Gets the CaseCloseReasons associated with the given theme",
-        responses={status.HTTP_200_OK: CaseCloseReasonSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="case-close-reasons",
-        methods=["get"],
-    )
-    def case_close_reasons(self, request, pk):
-        paginator = LimitOffsetPagination()
-        theme = self.get_object()
-        query_set = theme.caseclosereason_set.all()
-
-        context = paginator.paginate_queryset(query_set, request)
-        serializer = CaseCloseReasonSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Gets the CaseProjects associated with the given theme",
-        responses={status.HTTP_200_OK: CaseProjectSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="case-projects",
-        methods=["get"],
-    )
-    def case_projects(self, request, pk):
-        paginator = LimitOffsetPagination()
-        theme = self.get_object()
-        query_set = theme.caseproject_set.all()
-
-        context = paginator.paginate_queryset(query_set, request)
-        serializer = CaseProjectSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Gets the CaseCloseResult associated with the given theme",
-        responses={status.HTTP_200_OK: CaseCloseResultSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="case-close-results",
-        methods=["get"],
-    )
-    def case_close_results(self, request, pk):
-        paginator = LimitOffsetPagination()
-        theme = self.get_object()
-        query_set = theme.casecloseresult_set.all()
-
-        context = paginator.paginate_queryset(query_set, request)
-        serializer = CaseCloseResultSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-    @extend_schema(
-        description="Gets the subjects associated with the requested theme",
-        responses={status.HTTP_200_OK: SubjectSerializer(many=True)},
-    )
-    @action(
-        detail=True,
-        url_path="subjects",
-        methods=["get"],
-    )
-    def subjects(self, request, pk):
-        paginator = LimitOffsetPagination()
-        theme = self.get_object()
-        query_set = theme.subject_set.all()
-
-        context = paginator.paginate_queryset(query_set, request)
-        serializer = SubjectSerializer(context, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
-
-
-@user_passes_test(lambda u: u.is_superuser)
-def download_data(request):
-    DATABASE_USER = os.environ.get("DATABASE_USER")
-    DATABASE_PASSWORD = os.environ.get("DATABASE_PASSWORD")
-    DATABASE_HOST = os.environ.get("DATABASE_HOST")
-    DATABASE_NAME = os.environ.get("DATABASE_NAME")
-
-    filename = "zaken_db.sql"
-
-    command = f"PGPASSWORD='{DATABASE_PASSWORD}' pg_dump -U {DATABASE_USER} -d {DATABASE_NAME} -h {DATABASE_HOST} > {filename}"
-    os.system(command)
-
-    fl_path = "/app/"
-
-    fl = open(os.path.join(fl_path, filename), "r")
-    mime_type, _ = mimetypes.guess_type(fl_path)
-    response = HttpResponse(fl, content_type=mime_type)
-    response["Content-Disposition"] = "attachment; filename=%s" % filename
-    return response
 
 
 class ImportBWVCaseDataView(UserPassesTestMixin, FormView):
@@ -1184,19 +617,3 @@ class ImportBWVCaseDataView(UserPassesTestMixin, FormView):
             }
         )
         return self.render_to_response(context)
-
-
-class CaseCloseViewSet(
-    mixins.CreateModelMixin,
-    mixins.ListModelMixin,
-    mixins.RetrieveModelMixin,
-    viewsets.GenericViewSet,
-):
-    permission_classes = rest_permission_classes_for_top()
-    serializer_class = CaseCloseSerializer
-    queryset = CaseClose.objects.all()
-
-    def get_permissions(self):
-        if self.request.method not in SAFE_METHODS:
-            self.permission_classes.append(CanCloseCase)
-        return super(CaseCloseViewSet, self).get_permissions()
